requirements.txt
Flask
black
python-dotenv
django